python/requirements.txt
Flask==3.0.0
flasgger==0.9.7.1
flask-restful==0.3.10
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.6
python-dotenv==1.0.0
pytest==7.4.3